+++ packages/orq-rc/RELEASES.md
@@ -10879,3 +10879,13 @@ Based on:
 - [python v4.15.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.41 - packages/orq-rc
+
+## 2026-09-14 03:15:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.42 - packages/orq-rc